% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{generate_pie}
\alias{generate_pie}
\title{Creating a pie chart for imported transaction sheet}
\usage{
generate_pie(data, ...)
}
\arguments{
\item{data}{imported transaction sheet, should be stored as a data frame description}

\item{...}{currently ignored}
}
\description{
This function visualizes user's expenditure by category by creating a pie chart
}
\examples{
boa_data <- read_boa(read_example("boa_example_data.csv"))
generate_pie(boa_data)
}
