% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{read_boa}
\alias{read_boa}
\title{Reading transaction spreadsheet of Bank of American Accounts}
\usage{
read_boa(file, ...)
}
\arguments{
\item{file}{Name of the CSV file. It can be an absolute path, relative path, or a complete URL of the data file.}
}
\value{
Dataframe of a transaction data
}
\description{
This function reads transaction data from Bank of America accounts stored in a CSV file. It removes unnecessary
rows and columns, converts the date column to the Date format, categorizes transactions into expenditure or deposit, and assigns categories based on
transaction descriptions.
}
\examples{
read_boa(boa_data)
read_boa("boa_example_data.csv")
}
