% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{read_example}
\alias{read_example}
\title{Getting path to example data}
\usage{
read_example(file = NULL)
}
\arguments{
\item{file}{Name of example data file in string. Default is \code{NULL}, and it will list out the example files.}
}
\description{
There are 3 example transaction data files in \code{inst/extdata} directory of a perfi package.
This function returns a file path of the example data to utilize them.
}
\examples{
read_example()
read_example("boa_example_data.csv")
}
