% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{summary_stats}
\alias{summary_stats}
\title{Creating summary tables and visualizations of personal finance statistics}
\usage{
summary_stats(data, ...)
}
\description{
Creating summary tables and visualizations of personal finance statistics
}
