% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{budget_income}
\alias{budget_income}
\title{Returns income budgeted for user}
\usage{
budget_income(
  income,
  needs_percent = 0.5,
  wants_percent = 0.3,
  savings_percent = 0.2,
  freq = "bi-weekly"
)
}
\description{
Returns income budgeted for user
}
