% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{generate_pie}
\alias{generate_pie}
\title{Creating a pie chart for imported transaction sheet}
\usage{
generate_pie(data, ...)
}
\description{
This function visualizes user's expenditure by category by creating a pie chart
}
