% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfi.R
\name{read_boa}
\alias{read_boa}
\title{Reading transaction spreadsheet of Bank of American Accounts}
\usage{
read_boa(data_url, ...)
}
\description{
This function reads transaction data from Bank of America accounts stored in a CSV file located at the specified URL. It removes unnecessary
rows and columns,converts the date column to the Date format, categorizes transactions into expenditure or deposit, and assigns categories based on
transaction descriptions.Making it
}
